refactor: create Result component to handle error and success reponses, move alerts to hook
