+++ public/js/app/hooks/statisticsUseAsync.js
@@ -1,5 +1,7 @@
 import React, { useEffect } from 'react'
+import ReactDOM from 'react-dom'
 import { useWebContext } from '../context/WebContext'
+import { StatisticsAlert } from '../components/statistics/index'
 import fetchStatistics from './api/statisticsApi'
 
 const STATUS = {
@@ -65,11 +67,28 @@ function useAsync(asyncCallback, initialState) {
   return state
 }
 
+function renderAlertToTop(error = {}, languageIndex) {
+  const { errorType = '', errorExtraText = '' } = error
+  const alertContainer = document.getElementById('alert-placeholder')
+  if (alertContainer) {
+    ReactDOM.render(
+      <StatisticsAlert alertType={errorType} languageIndex={languageIndex}>
+        {errorExtraText}
+      </StatisticsAlert>,
+      alertContainer
+    )
+  }
+}
+function dismountTopAlert() {
+  const alertContainer = document.getElementById('alert-placeholder')
+  if (alertContainer) ReactDOM.unmountComponentAtNode(alertContainer)
+}
+
 function _getThisHost(thisHostBaseUrl) {
   return thisHostBaseUrl.slice(-1) === '/' ? thisHostBaseUrl.slice(0, -1) : thisHostBaseUrl
 }
 function useStatisticsAsync(chosenOptions) {
-  const [{ proxyPrefixPath, language }] = useWebContext()
+  const [{ proxyPrefixPath, language, languageIndex }] = useWebContext()
   const { documentType } = chosenOptions
   const asyncCallback = React.useCallback(() => {
     if (!documentType) return
@@ -82,6 +101,20 @@ function useStatisticsAsync(chosenOptions) {
   const initialStatus = { status: STATUS.idle }
 
   const state = useAsync(asyncCallback, initialStatus)
+
+  const { status: statisticsStatus, error = {} } = state || {}
+  const { errorType = '' } = error
+
+  useEffect(() => {
+    let isMounted = true
+    if (isMounted) {
+      if (errorType && errorType !== null) {
+        renderAlertToTop(error, languageIndex)
+      } else dismountTopAlert()
+    }
+    return () => (isMounted = false)
+  }, [statisticsStatus])
+
   return state
 }
